Rename mdacrawl.py to mdanalysis_core.py
